Splitting file.h into file.{h,c}
